requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-environ==0.9.0
django-filter==24.3
djangorestframework==3.15.2
environs==11.2.1
git-filter-repo==2.47.0
gunicorn==23.0.0
Markdown==3.7
marshmallow==3.23.3
packaging==24.2
psycopg2==2.9.6
python-dotenv==1.0.1
sqlparse==0.5.3
